backend/requirements.txt
streamlit
fastapi
uvicorn
python-multipart
opencv-python
numpy
pandas
scikit-learn
matplotlib
joblib
Pillow